nano/nano/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgi-correlation-id==4.2.0
black==23.7.0
click==8.1.5
fastapi==0.100.0
fastapi-mqtt==1.1.1
gmqtt==0.6.12
h11==0.14.0
httptools==0.6.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pydantic==1.10.11
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
requests==2.31.0
